+++ src/Symfony/Component/DependencyInjection/Loader/FileLoader.php
@@ -111,7 +111,9 @@ abstract class FileLoader extends BaseFileLoader
                 $r = $this->container->getReflectionClass($class);
                 $attribute = null;
                 foreach ($r->getAttributes(When::class) as $attribute) {
-                    if ($this->env === $attribute->newInstance()->env) {
+                    $attributeEnv = (array) $attribute->newInstance()->env;
+
+                    if (\in_array($this->env, $attributeEnv, true)) {
                         $attribute = null;
                         break;
                     }