+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/Prototype/Qux.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Component\DependencyInjection\Tests\Fixtures\Prototype;
+
+use Symfony\Component\DependencyInjection\Attribute\When;
+
+#[When(env: ['prod', 'dev'])]
+class Qux
+{
+}