+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/config/when_multiple_env.php
@@ -1,0 +1,7 @@
+<?php
+
+use Symfony\Component\DependencyInjection\Attribute\When;
+
+return #[When(env: ['test', 'prod'])] function () {
+    throw new RuntimeException('This code should not be run.');
+};