+++ src/Symfony/Component/DependencyInjection/Tests/Loader/PhpFileLoaderTest.php
@@ -182,4 +182,13 @@ class PhpFileLoaderTest extends TestCase
 
         $loader->load($fixtures.'/config/when_env.php');
     }
+
+    public function testWhenEnvAsArray()
+    {
+        $fixtures = realpath(__DIR__.'/../Fixtures');
+        $container = new ContainerBuilder();
+        $loader = new PhpFileLoader($container, new FileLocator(), 'dev', new ConfigBuilderGenerator(sys_get_temp_dir()));
+
+        $loader->load($fixtures.'/config/when_multiple_env.php');
+    }
 }